requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
yt-dlp>=2023.12.30
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiohttp>=3.9.0